[refactor] auth: 올바른 패키지 위치로 변경
