+++ src/main/java/org/example/ootoutfitoftoday/domain/auth/controller/AuthController.java
@@ -9,8 +9,12 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.ootoutfitoftoday.common.response.Response;
 import org.example.ootoutfitoftoday.domain.auth.dto.AuthUser;
-import org.example.ootoutfitoftoday.domain.auth.dto.request.*;
+import org.example.ootoutfitoftoday.domain.auth.dto.request.AuthLoginRequest;
+import org.example.ootoutfitoftoday.domain.auth.dto.request.AuthSignupRequest;
+import org.example.ootoutfitoftoday.domain.auth.dto.request.AuthWithdrawRequest;
+import org.example.ootoutfitoftoday.domain.auth.dto.request.RefreshTokenRequest;
 import org.example.ootoutfitoftoday.domain.auth.dto.response.AuthLoginResponse;
+import org.example.ootoutfitoftoday.domain.auth.dto.response.DeviceInfoResponse;
 import org.example.ootoutfitoftoday.domain.auth.exception.AuthSuccessCode;
 import org.example.ootoutfitoftoday.domain.auth.service.command.AuthCommandService;
 import org.example.ootoutfitoftoday.domain.auth.service.query.AuthQueryService;
